% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_occurrences.R
\name{clean_occurrences}
\alias{clean_occurrences}
\title{Nettoyer les occurrences spatiales des criquets pèlerins}
\usage{
clean_occurrences(df, lon_min = -20, lon_max = 65, lat_min = -10, lat_max = 40)
}
\arguments{
\item{df}{data.frame issu de \code{import_locust_data()}}

\item{lon_min}{Longitude minimale de la zone. Par défaut -20}

\item{lon_max}{Longitude maximale de la zone. Par défaut 65}

\item{lat_min}{Latitude minimale de la zone. Par défaut -10}

\item{lat_max}{Latitude maximale de la zone. Par défaut 40}
}
\value{
Un data.frame avec les occurrences nettoyées
}
\description{
Supprime les coordonnées invalides, les points marins,
les points en dehors de la zone Afrique/Maghreb/Moyen-Orient,
et les doublons spatiaux.
}
\examples{
\dontrun{
df_raw <- import_locust_data("occurrences.csv")
df_clean <- clean_occurrences(df_raw)
nrow(df_clean)
}

}
