% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate_data.R
\name{download_climate_data}
\alias{download_climate_data}
\title{Télécharger les données climatiques WorldClim}
\usage{
download_climate_data(
  var = "prec",
  res = 10,
  lon_min = -20,
  lon_max = 65,
  lat_min = -10,
  lat_max = 40,
  path = "data/climate"
)
}
\arguments{
\item{var}{Variable climatique : "tmax", "tmin", "tavg", "prec". Par défaut "prec"}

\item{res}{Résolution en minutes : 10, 5, 2.5. Par défaut 10}

\item{lon_min}{Longitude minimale de la zone. Par défaut -20}

\item{lon_max}{Longitude maximale de la zone. Par défaut 65}

\item{lat_min}{Latitude minimale de la zone. Par défaut -10}

\item{lat_max}{Latitude maximale de la zone. Par défaut 40}

\item{path}{Dossier de téléchargement. Par défaut "data/climate"}
}
\value{
Un objet SpatRaster (terra) avec les données climatiques
}
\description{
Télécharge les variables climatiques (température et précipitations)
depuis WorldClim et les découpe sur la zone d'étude des criquets.
}
\examples{
\dontrun{
clim <- download_climate_data(var = "prec", res = 10)
terra::plot(clim[[1]])
}

}
